Add support for overriding settings per request
This commit adds `res.settings`, an empty object on to which
properties can be defined that will override application
settings within Express iternals (the values from `app.get()`
are not altered.)

At present, this affects the following settings:
* view
* view engine
* views
* etag fn
* json replacer
* json spaces
* jsonp callback name

The `res.settings` object is passed through to view rendering code
by adding property `_settings` to the `options` parameter of
`app.render()`.

The entry key for the view cache was changed to prepend the `views`
setting to ensure the cache continues to work correctly if `views`
is overridden to a different value (eg if view `'users'` was cached
from `/projdir/views/a` but `views` was overridden to point to
`/projdir/views/b` for this request, it should not retrieve the
`'users'` view from the cache for `/projdir/views/a`).

`res.settings` does not set `__proto__: app.settings` since the
request can be routed to a sub-application which is not the same
as the initial application that initialises `req.settings`,
therefore `res.settings` only contains the values of the overridden
settings. An alternative implementation could try doing this, and
attempt to ensure `__proto__` was kept updated to the correct `app`
instance. This would be more convenient.